rename ab_classes.py, minor refactoring
